Set cftime library to a version > 1.1.0.
The NetCDF4 library uses the cftime library to convert dates.
With version 1.1.0, the default return type was changed to return a
cftime._cftime.DatetimeGregorian, which is not timezone aware. The two
new parameters only_use_cftime_datetimes and only_use_python_datetimes
enforces the return of a Python datetime object.